make dockerfile cache pip install requirements
since the `COPY` command that moves the source code into the container
isn't cached, we want to do as much setup (including installing reqs)
before that line of the dockerfile. So before we start on that, we
install all pip requirements. But then, to ensure that we run with
the required dependencies, after the COPY run pip install reqs.txt.

This second bit will capture any changed requirements, which we should
retroactively update the dockerfile with. For more info, see:
https://jpetazzo.github.io/2013/12/01/docker-python-pip-requirements/

Also, cleaned up the makefile and made sure some commands are private
since they're intended to only be ran inside the dockerfile (since they
don't invoke venv stuff)