Add endpoint to return count of pages in a letter
If we’re going to have the frontend fetch PNG previews of a letter one
page at a time then it will need to know how many pages are in a PDF.
Otherwise it will end up requesting pages that don’t exist. And while
try/catch is very Pythonic, it’s not very HTTP-omatic.

I think one page at a time is better than us rendering every page to
one image (which we could do) because it will give us more control over
how the pages are presented.

The `preview.json` URL isn’t the most self-expanatory, but it means
fewer changes to make on the admin side.