Add query parameter to specify which page to png
Because the frontend is going to be requesting each page of the PDF as
a separate PNGs it needs to be able to tell the preview app which page
it should send. A query parameter seems like a natural fit because it’s
filtering the view of the whole document.

This is kinda hard to test without reverse engineering the resultant
PNG. We don’t do any testing like that in this repo currently, so I’m
not about to start here. Can probably assume that if it `400`s when
expected then a `200` is good.