Build libcairo by hand
The version of libcairo that comes with Debian stable is quite old. It
gives some warnings about ‘known rendering problems’, which might be
why logos in our image previews look a bit crusty.

This commit adds step to the `Dockerfile` to build Cairo from source,
rather than use Apt.

N.B. I have not yet found a way to check if our Python apps are using
this version, or still using the old one. So, while this does run,
consider it a WIP.