Use raster file format for logos in letter preview
Initially we thought that a vector image would be best, because then we
don’t have to worry about resolution.

But the library that turns our PDF letters into PNGs seems to have a
hard time rasterising SVGs without making a mess of them. So let’s make
its life easier and give it a pre-rasterised file.