Tie the CMYK transformation into /print.pdf
This commit makes the /print.pdf endpoint return a PDF with the text
converted to CMYK colour.

It does this by:

- generating the PDF as before (using Weasyprint)
- feeding the pdf into PDFTK, a command line programme that can
  decompress PDF data into PDF code
- going through each line of the PDF looking for colours to replace
- feeding the replaced PDF code into GhostScript, which writes it back
  into compressed PDF data

This lets us produce PDFs with text that is 100% K ink, rather than
a mixture of C, M, Y and K, which is what happens with a naïve
conversion from RGB to CMYK. Pure black text is better because it avoids
the fuzzy edges that result from overlapping multiple ink colours.