Add optional CMYK logos for printing letters
Letters will look best printed with a CMYK, vector logo. Previews will
still look best by using an RGB, raster version of the logo.

We have good-looking RGB logos already using PNGs.

This commit adds the option to specify a version of the logo for
printing. Experimentation has found that a CMYK SVG works well, and has
its colour space preserved in the outputted PDF.

This commit implements the CMYK version of the HM Government logo only,
but refactors the code to make it easy to add more CMYK logos in the
future.

Eventually this should be a combination of things stored in the database
and logo images stored in S3, so that we can manage logos via platform
admin. For expediency we’re going to continue managing these manually
for now.