Use RGB vector logos for print, then convert
Turns out:
- SVG is the only vector format that works in web pages (which is what
  our letters are generated from)
- SVG does not support CMYK except in version 2, and there are no
  graphics tools which support SVG version 2

So this commit changes things a bit. It keeps:
- support for a new, SVG, version of the logo for printing

It removes:
- the implication that this vector version is CMYK

Luckily, the CMYK conversion stuff we wrote before is able to target the
colours in the vector logo. So we just need to make sure we have the
colour conversion set up properly to properly map any colours that we
have in logos. This is hard coded for now, but we can probably make it
more intelligent in the future.