Added a log handler to template-preview so that statsd messages can be monitored
- Whilst testing the statsd changes it was apparent that the stats would
be send without a prefix and appear as 'None' in Grafana. Added the
config that the statsd client uses in utils to set the application name
and the environment. The sqs_queue_prefix is used so that 'live' is used
instead of 'production' (because its used else where in grfana as 'live')
- Added a handler for the logger so that messages can be printed to the
log at level but ERROR. Gunicorn prints are stderr so this allows messages
to be printed to stdout. http://docs.gunicorn.org/en/stable/settings.html