Use common format for specifying colour values
It’s unusual to specify colour values as numbers from 0.0–1.0. This
means manually converting to this format when adding new rows to the
colour mapping file.

Most graphics software uses values from:
- 0–255 for RGB triplets
- 0–100% for CMYK quadruplets

By programmatically converting from these formats to the decimal values
needed for the transformation we reduce the risk of introducing error
by manual conversion, and make it quicker to add new values to the
mapping.