Add single identity colours to mapping
There’s a fair chance that we’ll one day be sending letters with logos
of most government departments. So we should make sure that we get their
colours accurate.

The branding for most central government organisations is defined by
the ‘single identity’. This commit adds those colours to the mapping.
There are other, non single identity, logos which we will also have to
deal with later, either by adding to the mapping or writing some code
which takes a best guess at the transformation.

The colour values added here are eyedropped from the document
‘Government Identity System CMYK.pdf’ which can be found in the
‘__AssetLibrary-Design’ folder on Google Drive.