Updated the creation of the unique name to handle if there is personalisation
This is to cache the variants of the personalisation with a new unique
name as each different set of values can be different. This still has
a benefit as the preview creates the pdf then each page of the png,
meaning this method is called upto four times per preview, three of which
would recreate the pdf data which takes around 300ms.