Changed the redis key name to be a unique dict based on the request
Changed the key name which was based on the created date to be based on
a dict formed out of the unique attributes of the request, specifically
the template and the personalisation.

Removed the need for another endpoint and replaced the functionality of
the current endpoint and therefore the associated extra tests.