Ensure page counts works if PDF comes from cache
For some reason the IO objects that Boto returns aren’t iterable
(whereas a normal `BytesIO` is). We were dealing with this in the
endpoints that actually serve the PDF by calling `.read()` on the
object. But we weren’t doing the same handling in the endpoint that
takes the PDF and returns the page count.

This commit makes sure:
- adds new mocking to replicate the behaviour of Boto
- makes sure that the page count endpoint calls `.read()` on the object
  it receives (which could be from the cache or a generated PDF)