Don’t hash page no. in precompiled PNG cache key
This will make it easier to debug because we’ll be able to match which
PNGs were generated from the same letter.

We’re already doing this for templated letters; this makes it consistent
for precompiled letters.

This commit also namespaces precompiled and templated letters to make it
clearer what’s going on for someone looking in the bucket.