After discussions with the DVLA the right border is less important than the left border, as the left border is where the bar codes etc go. The current DWP letter goes through the system would fail the validation check because it has some text which encroaches in the right border however we know it is good.
Validation changed so the right border is not checked but the overlay
returns the hard border i.e. including the right border. This is likely
to be used for a tool to help with adoptions. Also added a query string
parameter to allow developers to get the same png which the validation
is based on to help with debug i.e. if the validation fails use curl
or postman to get the fails validation image.

* Updated width of the border if the overlay parameter is passed into the
function
* Added a query string parameter to allow developers to call the same
overlay endpoint but get the validation png not the png intended for the
tool