only add notify tag if the notify tag isn't already present
if the pdf doesn't already contain the word NOTIFY, we should add it.
But if it does, we don't want to add it twice - we're not sure how the
printers will deal with that.

we can't guarantee that the notify tag is both present AND correct,
but if it's there we should assume that it's already good, since the
creators of the PDF know about it.