Host letter images on s3
When hosting locally, each request to generate a pdf will also make
several more requests to the template preview box, for each static
image - any logo, and also, if rendering a preview for the admin app,
then qr code and barcode placeholders etc. If there are five concurrent
requests, then they'll use up all of the workers and deadlock each
other.

Moving the images to the `public-logos` buckets (currently used for the
email logos) gets rid of this deadlock. Also it helps us move away from
defining all of the images in a github repo.