return 502 if images (logos) aren't available when rendering letters
Logos are hosted on s3. Previously, they were committed to the repo and
we had unit tests asserting they were all present - but now there's
nothing checking that they are actually valid before rendering. If an
image doesn't return then weasyprint supresses the exception. This will
result in letters being sent out with no logos - pretty disastrous.
Weasyprint doesn't offer a way to override this - you can pass in a
custom url_fetcher, but it still wraps the whole thing in a try except.

After it catches the exception, it does log an error, so as a hack, we
monkeypatch the weasyprint logger object to call abort with a 502 if
the expected exception is given.