Test the behaviour of 'convert_to_cmyk' on CMYK and RGB input
This commit serves to document manual tests on 'convert_to_CMYK'
function.

What was done:

Two pdf files were provided to the function, one contained
an image in RGB colour mode and the other contained the same image
in CMYK colour mode. Then outputs have been saved to png files,
extracted from the docker container using:
'cp <container-number>:/home/vcap/app/<path-to-file-within-app> .'
And visually compared for differences. Visually, the structure of
the output images was the same and displayed colours looked
almost the same, if not identical.

Later, third test was created to programmatically compare the two
files, and this test found that the files were not 100% identical.
I later understood that this was because the RGB input
and the CMYK input had different margins/cropping.

Still, since the files bear no visible differences, and no unexpected
changes to CMYK input were made, it was decided that it is ok to
use 'convert_to_cmyk' method on CMYK input, as well as RGB input.