Check if pdf already contains CMYK images before converting
a precompiled letter to CMYK.

As per the previous commit,the output of
convert_pdf_to_cmyk function was visually very similar for
the same image provided as rgb and cmyk image. Still, this did not
100% guarantee the outcome. What if an image we were processing
was CMYK-encoded in a different way to our test file and the
printerd result would be unsatisfying?

Hence, a new function has been added if a PDF already contains
CMYK images, and the PDF only gets converted to CMYK if it does not
contain pre-existing CMYK images.

This approach also has drawbacks: if a single PDF has both
CMYK and RGB images, then it will not get converted to CMYK.