Reverse validate flag logic in /precompiled/overlay.png
`validate=false` should mean we want to generate a preview with a
red overlay instead of an empty PDF with pixels in non-printable
areas, but was doing the opposite in the existing code.

Since we want this behaviour by default (without any params) we
explicitly check for `validate=true` instead of `false`, so when
the argument isn't set we get a semi-transparent overlay.