All-page validate and preview for pre-compiled letters
For the new admin endpoint to validate boundaries of pre-compiled
letter PDFs we need to display all PNG pages and any validation
errors on one page. In order to avoid storing the PDF and making
multiple separate requests to template-preview we add a new flag
to the validate route that returns base64-encoded PNG images of
all PDF pages that could be included into the admin app page using
data-uris.

As a part of this we:
- added a preview flag to validate precompiled endpoint
- with flag, it returns preview pngs, with overlay if
validation failed, and without overlay otherwise
- upgraded overlay_template_areas to handle multiple pages pdf
- created pdf_to_pngs function that converts multiple-pages
pdf to pngs and returns a list of png pages