Disable address rewrite for pre-compiled letters
We don't know whether address rewriting produces the expected
result on actual pre-compiled letters since we don't have enough
example letters to check that it doesn't break the address block
on different PDFs. We're disabling it for now with the intention
to turn it back on once we've manually checked it's parsing the
correct address from the variety of letter templates.