Use git commit SHA as the Docker image tag
Due to how CloudFoundry stages docker containers deployed apps will
scale using the latest container version tagged with a given tag.

This means that Docker tags need to be immutable in order to avoid
accidentally starting up instances with a failed release code.

Changing the image tag to use a short git commit SHA (and a `-dirty`
suffix for builds with uncommitted changes) makes sure that each
release gets it's own separate tag, which makes it easier to rollback
and deploy the app locally.