Move docker image to a Docker Hub organization
govuknotify is a user account that is used by Jenkins to push images
to Docker Hub.

Switching images to an organization allows us to give permissions
to push containers from dev machines without having to switch to
Jenkins user credentials.

Jenkins user account is an owner of the organization, so it should
be able to push images without any changes to our CI configuration.