Build latest versions of ghostscript and poppler
Upgrading gs to 9.25 solves the problem with RGB-to-CMYK pre-compiled
PDF conversion creating an invalid PDF on some files.

We're running into a number of issues with our PDF/PNG generation
related to version of the underlying tools we're using.
Relying on system packages makes it hard to upgrade/downgrade these
tools to a specific version since there's usually only one available
in the system's package repository.

As a partial workaround for this we used to use a mix of Debian Jessie
and Stretch repositories to install package versions that were close
enough to what we wanted. This approach still left us with few choices
of available versions (most of them not up-to-date) and made apt hard
to work with (since it now required explicitly choosing the Debian
release).

Instead, this switches our base container to the latest stable Debian
release (Stretch) and compiles the required versions of the tools we
rely on after installing all the build dependencies.

This increases the time it takes to build the container from scratch
and can potentially lead to conflicts with system packages (but this
hopefully shouldn't be a big problem since not a lot of things rely
on pdf libraries), but it gives us direct control over gs/poppler
versions.