Use letter filename if available instead of dvla_org_id
In the preview_schema, dvla_org_id is now optional instead of required.

For POST endpoints which were expecting to receive the dvla_org_id we
now check for a letter filename in the data first and only use the
dvla_org_id if a filename is not provided. notifications-api and
notifications-admin are already passing the filename through, so this
commit ensures that we now start using this value.