Use the Jenkins build number as Docker tag when promoting a release
Since our deploy jobs don't have a git checkout available we can't
use the commit SHA as the tag unless we pass it in as a variable.

Instead, we use the original master branch build job as the Docker
image tag.

This also removes PaaS artifact targets that aren't useful for an
app deployed with a Docker image.