downgrade ghostscript to 9.21
we use ghostscript to, among other things, convert our PDFs from RGB to
CMYK. When we upgraded ghostscript to 9.22, something changed and the
grayscale colours it output are no longer srgba(0, 0, 0, x), but are
srgba(7, 5, 5, x). (When looking at output of `identify -verbose`).
DVLA can’t convert these to true blacks when printing, so everything
will be fuzzy and grey.

Pinning to 9.21 fixes this, but also means we are conciously ignoring
this CVE: https://nvd.nist.gov/vuln/detail/CVE-2018-11645

But that's okay because it only allows people to find out what files
are on disk and it's a public docker image anyway 🤷