[Spike] Generate PNG versions of logos from SVG
Currently we do this manually. We can make our process more efficient
by automating it.

This also has the benefit of being able to reuse the caching code
already in template preview, so we can generate the PNG on the fly then
cache it for next time, and we don’t have to write extra code to put the
converted file back to S3.