Adjust position of white Notify text
The hidden NOTIFY text tells the printer when a new letter starts. It
has trouble finding the text if it overlaps any other elements on the
page. This would be impossible if the NOTIFY text was fully in the
non-printable area. However the right-hand-edge of the text was
intruding in the printable area by a fraction of a millimeter.

This commit moves the NOTIFY text to be in the center of the
non-printable left margin, which should make it well clear of anything
on the page.

This means DVLA can change the boundaries of where they’re looking
for the NOTIFY text. Until then their existing code should work before
and after this change (to be confirmed – do not deploy until it is).