rewrite address blocks.
we've had a couple of problems with non-conforming address blocks
lately, misaligned and encroaching in to the MDI return address
block at the top. While we've changed the validation to stop this, we
should also look at bringing back the address sanitisation to make sure
we don't get this problem in future.

To recap, this will take all text in the address block, split it into
lines, and then cover it up with a white bg and re-write it in 8pt
arial with correct line spacing etc.