Completely redact address block of a precompiled letter
Previously, we were using white rectangle to cover the address block
in precompiled letters sent by our users. Then we put properly formatted
address block on top of the white rectangle.
Unfortunately, the old address block was still machine-readable from
below the white rectangle, and it sometimes confused DVLA machines
that read postcodes and sort letters for sending.

Hence, we have to ensure that the address block is totally redacted
out of existence. This PR does that, with help of code from
[pdf_redactor](https://github.com/JoshData/pdf-redactor) github repository.

We now:
- find the address block in pdf's text stream and delete it.
- if we find more than one match, we do not redact, as this could
remove information form letter content block, and we log that we
found multiple matches for the address block.
- we also log if we did not find a match.
- we still use the white rectangle, in case one of the above
happens and hence we are unable to perform full redaction.