move ValidationFailed error handler to generic precompiled one
If validation fails unexpectedly, we should handle this the same as if
it cannot open the PDF.

the /precompiled/ blueprint currently has four endpoints:

* /precompiled/sanitise
* /precompiled/add_tag
* /precompiled/validate
* /precompiled/overlay.<file_type>

add_tag isn't used and validate is going away soon (when the one-off
pdf epic is shipped), so we don't need to worry about them. overlay is
used to display images that have already been through the validation
process so we know they are at least processable.

So really, sanitise is the only endpoint that looks at PDFs for the
first time, running the risk of being unable to open them. We've seen
errors with PyPDF2 where it raised a NoneTypeError rather than a
regular validation error, so we're now catching all `Exception` errors
in a blueprint level handler (so it doesn't interfere with exceptions
in other endpoints).

It looks for `page_count`, `message` and `code` attributes on
exceptions, but otherwise returns the same json as the existing
ValidationFailed error handler.