Use command in manifest instead of Procfile
We were attempting to get the Procfile to use exec so that signals are
propagated to the app correctly.

However when prepending "exec" in front of the start command in the
Procfile, it did not have an affect.

It is hard to reason about how Docker applications use the Procfile, so
we thought it would be better to explicitly set the command in the
manifest.

Additionally we have made the manifest look more similar to other notify
manifests (taken from notifications-ftp)

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>
Co-authored-by: Becca <rebecca.law@digital.cabinet.office.gov.uk>
Co-authored-by: Toby <toby.lornewelch-richards@digital.cabinet.office.gov.uk>