Change default CF_SPACE to be development, not sandbox
We set the `NOTIFY_ENVIRONMENT` to be the same as the `CF_SPACE`, which
defaulted to `sandbox`. Since we want to prefix bucket names and queues
with `development`, this changes the default of `CF_SPACE` to be
`development`.

Also removed the `sandbox` target from the Makefile, since we don't
deploy to Sandbox.