Remove unused .cfignore
We deploy this app using docker and so we don't push any files directly
from the repository when we do a cf push. Therefore the cfignore file is
redundant.

I've tested this by checking that files in this repository that are
not ignored by the gitignore but aren't included in the docker image are
not found in /home/vcap/app on this app running in preview.