Remove references to build numbers
We are passing several args into the docker container that are not being
used anywhere. Some variables such as build_number are no longer
relevant because we don't really care about the build number or
build URL as people can just go and find it in concourse if they need
to.