fix lack of commit sha in status page
Was failing because there was no .git folder to query to find out the
commit sha. have added that into our production images. It's not ideal
to add the .git repo into the container for prod as it should contain as
little as possible it needs to run the application, however this was a
quick solution and not worth too much thought.