Refactor sanitised endpoint
- Refactored the sanitise endpoint, to pull most of the code out into a function
that can be used by both the existing endpoint and the task that will be created.
- Removed the precompiled blueprint error handler and handled errors in the new
sanitise function instead. This means that the `validate_pdf_document` endpoint
won't have its errors handled in the same way (so have updated a test) but this
endpoint is deprecated and will be removed.
- Added a for the situation where sanitising a letter throws an unexpected error.