Added a new sanitise_and_upload_letter task
This task gets precompiled letters which have passed virus scan from the
scan bucket and attempts to sanitise them. If the file was valid, the
sanitised version gets uploaded to the new template preview sanitise bucket.
The `process-sanitised-letter` task in notifications-api with is then
always called with the result and some other details.

If there was a Boto error, we don't want to retry the task (since it may
be that the letter has already been through all the sanitise tasks), so
we just log an exception.