Use Celery 4
Switched to using Celery 4 which means that we can git rid of our
patched version of Boto. Pycurl, which is included in celery[sqs] needed
`libcurl4-openssl-dev` and `libssl-dev` to be installed in order for it
to work.

We now long polling with a wait time of 20 seconds.

Have updated the config names to use the new Celery 4 format and put
them under the `celery` key in the config dict to keep things clear.
Also removed Celery config that we were setting to the default config,
as this was having no effect.