Error in the Makefile if NOTIFICATION_QUEUE_PREFIX unspecified
Using a function to get the queue prefix and raise an exception if
`NOTIFICATION_QUEUE_PREFIX` wasn't set was causing an error to be raised
when running the Flask app. The message about the queue prefix not being
set was quite far up the stack trace, so it wasn't obvious what the
issue was.

This change moves the error to the Makefile when running the app using
`make run-celery-with-docker` (which is the only way that the app will
be run). It's now not a problem to not have `NOTIFICATION_QUEUE_PREFIX`
When running the Flask app.