make all margins 1mm more generous
when we validate, we use margins provided by DVLA to validate against to
prevent people writing in places where either

* machines can't print on the edges of the paper
* DVLA will put barcodes and things that we shouldn't cover up

However we don't need to be as precise on these borders. Utimately, a
few pixels just over a border will be fine. This might happen because
someone's aligned text, but the text goes outside the normal bounding
box (eg the top left of a letter W). Also, when we check boundaries we
convert to an image and then check the pixel colours - this may cause
issues if anti-aliasing puts a fuzzy grey pixel just over.

Add a 1mm margin-of-error to each border so that if someone puts a
letter right on the edge, it won't trip our validation. This change also
means that if the validation does fail, the red boundaries they see on
their failed pdf will also be pushed out 1mm. However, we're
deliberately not updating the guidance template so that they still
target the original DVLA margins.