move from x, y, width, height to x1, y1, x2, y2 when drawing on pdfs
All of our pdf drawing stuff (blocking out areas for validation or
highlighting the borders in red) comes from our letter spec [1]. Our
letter spec defines virtually every measurement from the top left of the
page, and lists all of them in mm.

In contrast, the `canvas.rect` takes an x and a y value from the bottom
left (as that is how pdfs are defined internally), and a width and a
height. It also requires these values in points (1/72th of an inch).

By wrapping the canvas.rect call, we can take care of this abstraction
in one place. Rather than lots of complicated unit conversions and
width/height calculations, we can just deal exclusively in pairs of
coordinates, which is a bit easier to reason about.

[1] (https://docs.notifications.service.gov.uk/documentation/images/notify-pdf-letter-spec-v2.4.pdf)