make _colour_no_print_areas_in_red only expect one page
previously i was confused and thought we passed the entire PDF in. In
reality, the admin app passes in a single page pdf that it has extracted
from the full thing.

since we expect just a single page PDF, put in an assert that validates
that, and also, we can remove the for loop in the _colour function and
change the page_number to just be an `is_first_page` boolean.