re-introduce /precompiled/overlay.pdf endpoint
this endpoint is used by the notifications endpoint to view a
notification. That gives you the option to download a PDF that was
submitted precompiled via API but failed validation. The PDF is passed
in, and template-preview takes the PDF, adds the red borders to it, and
returns it.

this commit refactors the `_colour_no_print_areas_of_page_in_red`
function to take a `PageObject` as returned by PdfFileReader.getPage,
and modifies that page in place. This makes it easier to call
successively for every page in a pdf