add code to check embedded fonts
We have PDFs that DVLA can't print automatically - we suspect it's due
to unembedded fonts. As a reminder, fonts can optionally be embedded in
a pdf - if they're not, whatever is accessing the PDF needs to have that
font locally for it to load properly. We think that making sure all
fonts are embedded will be enough to make these files printable.

using code from https://gist.github.com/tiarno/8a2995e70cee42f01e79.
This code comes up with a few false positives (it thinks a font is not
embedded, but adobe acrobat says the font is embedded), but no false
negatives as far as I can tell.