split up the sanitise function and use BytesIO everywhere
split up the sanitise function - flake8 was whinging, with fair cause,
that the function was too big. Separated the functions that modify the
PDF from the rest of the logic around building up the json and checking
intiial validity etc.

also use BytesIO everywhere. In functions that modify, don't worry about
seeking since the old data will probably be discarded. But in functions
that you wouldn't expect to modify the pdf (such as colourspace or font
checking), then seek(0) at the end.

be kind, rewind!