redact pdfs with fonts that extend WinAnsi or MacRoman encoding
Welcome to the exiting world of PDF fonts!

The PDF Redaction tool works in unicode, so that we can redact text that
is unicode. The actual text is stored in binary within the pdf (as in a
b'string'), and so pdf_redactor.py has to figure out what font that data
is written in, and then decode it into unicode. Then, at the end it
encodes it back into the original encoding.

Seems reasonable. However, pdf_redactor looks for a few things. The font
needs a ToUnicode table, or to be one of WinAnsi or MacRoman encoding.
If the font has none of these, then pdf_redactor just treats the text as
a "?", so as a result it won't be redacted.

We have seen, in real letters, fonts that look like the following.

{
    '/Type': '/Font',
    '/Subtype': '/TrueType',
    '/BaseFont': '/Arial',
    '/Encoding': {
        '/Type': '/Encoding',
        '/Differences': [
            '128',
            '/Euro'
        ],
        '/BaseEncoding': '/WinAnsiEncoding'
    },
    '/FirstChar': '30',
    '/LastChar': '255',
    '/Widths': (9, 0),
    '/FontDescriptor': (10, 0)
}

Here, rather than the /Encoding object being a string, it's a sub-object
that lists a base encoding, and then a list of differences[1]. In the
above example, it's WinAnsiEncoding, except for it replaces `chr(128)`
with the euro symbol. (WinAnsiEncoding is `cp1252` in python). Now this
is confusing, because WinAnsiEncoding already has the € symbol at
128[2], but who knows, I'm chalking this one up to arbitrary PDF
bullshit as per usual.

Since this font is essentially identical to regular WinAnsi, I think
it's safe to use the base encoding. Worst case: `string.encoding`
crashes, we get a technical failure, an alert, and can investigate. My
hunch is that that isn't especially common, but if it is, we could
always work around it later.

refs:

[1] https://blog.idrsolutions.com/2011/04/understanding-the-pdf-file-format-%E2%80%93-custom-font-encodings/
[2] `chr(128) == '\x80'. b'\x80'.decode('cp1252') == '€'`