don't overwrite addresses if we can't redact them
previously, if we were unable to remove the underlying address (this is
because it either can't be found programatically due to pdf weirdness,
or because the address is repeated elsewhere on the first page), then we
would put a white box over the old address and re-write it in the
correct position/font/etc.

However, DVLA have issues with their automated tooling picking up both
the new address and the old one, as it's just hidden, not removed.

If we can't redact the old address, just leave the PDF as it is. This
may mean that the font isn't correct and position isn't perfect, but as
the alternative involves DVLA manually printing, this is the best we can
do.

Also make some more of the code BytesIO Everywhere.