use vector version of logos when rendering templated previews
When a platform admin uploads a new svg logo, we generate a PDF, and
store both the original svg and the pdf on s3.

When someone previews a PDF on the admin app, we take the converted png,
embed it in an HTML document, convert that html to a PDF, then convert
each page of that document to a png.

When we create the letter to send to DVLA, we take the original svg,
embed it in an HTML document, convert that html to a PDF, then send that
PDF to DVLA.

We previously had these separated because we ran into issues with the
pdf -> png conversion when the pdf had SVGs. Now that we generate PDFs
ourselves rather than sending markup to DVLA, and we've upgraded the
underlying libraries (especially cairo and weasyprint), we don't need to
embed pngs - embedded SVGs work just fine. This means that the preview
png is much closer to the actual PDF we send to DVLA, improving our
confidence that we won't crop logos.

I've kept the PNG conversion stuff in for now - we can get rid of it in
a later commit. But it's no longer being used.