Ignore Celery logs when running on Paas
We want to ignore the logs that come from Celery itself since we
already have logs that cover the tasks - the Celery logs provide
duplicate data. We have already redirected the Celery logs to /dev/null
when running in Docker, but this add the same behaviour when running
on the Paas.