Unset the statsd_prefix for template preview
As mentioned in https://statsd.readthedocs.io/en/v2.1.2/configure.html
we could set a statsd prefix to help statsd identify where our statsd
metric is coming from. However this is already something we are doing in
the statsd client using a `namespace` variable that ensures all metrics
start with `environment_appname`. See
https://github.com/alphagov/notifications-utils/blob/master/notifications_utils/clients/statsd/statsd_client.py#L41

Therefore we are currently sending statsd metrics of the form
`production_production_notifications_template_preview_template_preview__generate_png_page`

The first `production` is redundant here and this PR will remove it for
all environments and bring this app in line with all our other metrics.

Note, on the downside, our statsd to prometheus mapping is not picking
up existing template preview statsd metrics correctly. I don't know whether these metrics from
template preview will now be parsed correctly or not, but either way
this is a step in the correct direction and some point in the future we
should have a review of all the statsd metrics anyway.
https://github.com/alphagov/notifications-aws/blob/master/paas/statsd/statsd-mapping.yml